requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==1.11.20
django-anymail==6.0
django-bootstrap4==0.0.8
django-filter==1.1.0
django-registration-redux==2.5
idna==2.8
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
